Delete route security with authentification

+++ src/user/user.resolver.ts
@@ -1,8 +1,10 @@
 import { Resolver, Query, Args, Mutation } from '@nestjs/graphql';
 import { UserService } from './user.service';
 import { UpdateUserDTO } from './dto/update-user.dto';
-import { UseGuards } from '@nestjs/common';
+import { UseGuards, UnauthorizedException } from '@nestjs/common';
 import { GraphqlAuthGuard } from '../guards/gql-auth.guard';
+import { CurrentUser } from '../decorators/user.decorator';
+import { User } from '../types/user';
 
 @UseGuards(GraphqlAuthGuard)
 @Resolver()
@@ -22,8 +24,12 @@ export class UserResolver {
   // TODO: Implement who am i
 
   @Mutation()
-  async delete(@Args('email') email: string) {
-    return await this.userService.deleteUserByEmail(email);
+  async delete(@Args('email') email: string, @CurrentUser() user: User) {
+    if (email === user.email) {
+      return await this.userService.deleteUserByEmail(email);
+    } else {
+      throw new UnauthorizedException();
+    }
   }
 
   @Mutation()
